+++ src/sensors/hall_effect/mod.rs
@@ -3,6 +3,10 @@
  * Copyright 2025 Sira Pornsiriprasert <code@psira.me>
  */
 
+mod hall_latch;
+mod hall_switch;
 mod linear_hall_sensor;
 
+pub use hall_latch::HallLatch;
+pub use hall_switch::HallSwitch;
 pub use linear_hall_sensor::LinearHallSensor;